todo funcional

+++ UT4/02-sessions/05-linkenin/public/logout.php
@@ -3,18 +3,18 @@
 require('../src/init.php');
 
 if(isset($_SESSION['usuario'])) {
+    
     unset($_SESSION['usuario']);
     unset($_SESSION['id']);
 
-    $DB->ejecuta("DELETE FROM token WHERE valor = ?", $_COOKIE['recuerdame']);
-
-    // unset($_COOKIE['recuerdame']);
-    // $_COOKIE['recuerdame'] == null;
-    setcookie("recuerdame", null, [
-        "expires" => time() - 3600,
-        // "secure" => true,
-        "httponly" => true
-    ]);
+    if (isset($_COOKIE['recuerdame'])) {
+        $DB->ejecuta("DELETE FROM token WHERE valor = ?", $_COOKIE['recuerdame']);
+        setcookie("recuerdame", null, [
+            "expires" => time() - 3600,
+            // "secure" => true,
+            "httponly" => true
+        ]);
+    }
 }
 
 header('Location: listado.php');
